fix rounding issue

+++ examples/src/main/kotlin/actor/proto/examples/inprocessbenchmark/InProcessBenchmark.kt
@@ -46,7 +46,7 @@ fun run() {
 
         val elapsedMillis = (currentTimeMillis() - sw).toDouble()
         val totalMessages = messageCount * 2 * clientCount
-        val x = ((totalMessages.toDouble() / elapsedMillis * 1000.0).toInt())
+        val x = ((totalMessages.toDouble() / elapsedMillis.toDouble() * 1000.0).toInt())
         println("$t\t\t\t\t$elapsedMillis\t\t$x")
         for ((client, echo) in pairs) {
             client.stop()
